src/main/java/com/nicekkong/bootservice/webservice/service/PostsService.java
package com.nicekkong.bootservice.webservice.service;

import com.nicekkong.bootservice.webservice.dto.PostsSaveRequestDto;
import com.nicekkong.bootservice.webservice.domain.PostsRepository;
import com.nicekkong.bootservice.webservice.dto.PostsMainResponseDto;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.List;
import java.util.stream.Collectors;

@AllArgsConstructor
@Service
public class PostsService {

    private PostsRepository postsRepository;

    @Transactional
    public Long save(PostsSaveRequestDto dto) {
        return postsRepository.save(dto.toEntity()).getId();
    }

    @Transactional
    public List<PostsMainResponseDto> findAllDesc() {
        return postsRepository.findAllDesc()
                .map(PostsMainResponseDto::new)
                .collect(Collectors.toList());
    }

}
